Add peripheral interface

+++ Core/Inc/Peripherals/Neopixel/neopixel.h
@@ -8,11 +8,12 @@
 #ifndef INC_PERIPHERALS_NEOPIXEL_NEOPIXEL_H_
 #define INC_PERIPHERALS_NEOPIXEL_NEOPIXEL_H_
 
+#include <stdint.h>
 
 #define NEOPIXEL_NUMBER 23
 
 
-void Set_Neopixel(int LEDnum, int Red, int Green, int Blue);
+void Set_Neopixel(uint32_t LEDnum, uint8_t Red, uint8_t Green, uint8_t Blue);
 
 void Neopixel_Send(void);
